Expose the usage searching function via RegexSearcher
