+++ crates/maple_cli/src/command/dumb_jump.rs
@@ -9,122 +9,9 @@ use anyhow::Result;
 use rayon::prelude::*;
 use structopt::StructOpt;
 
-use crate::dumb_analyzer::{
-    definitions_and_references, definitions_and_references_lines, find_occurrence_matches_by_ext,
-    get_comments_by_ext, get_language_by_ext, MatchKind,
-};
+use crate::dumb_analyzer::{RegexSearcher, UsagesInfo};
 use crate::tools::ripgrep::{Match, Word};
 use crate::utils::ExactOrInverseTerms;
-
-/// All the lines as well as their match indices that can be sent to the vim side directly.
-#[derive(Clone, Debug, Default)]
-pub struct Lines {
-    pub lines: Vec<String>,
-    pub indices: Vec<Vec<usize>>,
-}
-
-impl Lines {
-    /// Constructs a new instance of [`Lines`].
-    pub fn new(lines: Vec<String>, indices: Vec<Vec<usize>>) -> Self {
-        Self { lines, indices }
-    }
-
-    /// Prints the lines info to stdout.
-    pub fn print(&self) {
-        let total = self.lines.len();
-        let Self { lines, indices } = self;
-        utility::println_json_with_length!(total, lines, indices);
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.lines.is_empty()
-    }
-}
-
-// TODO: a new renderer for dumb jump
-fn render(matches: Vec<Match>, kind: &MatchKind, word: &Word) -> Vec<(String, Vec<usize>)> {
-    let mut group_refs = HashMap::new();
-
-    // references are these occurrences not in the definitions.
-    for line in matches.iter() {
-        let group = group_refs.entry(line.path()).or_insert_with(Vec::new);
-        group.push(line);
-    }
-
-    let mut kind_inserted = false;
-
-    group_refs
-        .values()
-        .flat_map(|lines| {
-            let mut inner_group: Vec<(String, Vec<usize>)> = Vec::with_capacity(lines.len() + 1);
-
-            if !kind_inserted {
-                inner_group.push((format!("[{}]", kind), vec![]));
-                kind_inserted = true;
-            }
-
-            inner_group.push((format!("  {} [{}]", lines[0].path(), lines.len()), vec![]));
-
-            inner_group.extend(lines.iter().map(|line| line.build_jump_line_bare(word)));
-
-            inner_group
-        })
-        .collect()
-}
-
-struct RegexSearcher {
-    word: String,
-    extension: String,
-    dir: Option<PathBuf>,
-}
-
-impl RegexSearcher {
-    async fn search_references_or_occurrences(
-        self,
-        classify: bool,
-        exact_or_inverse_terms: &ExactOrInverseTerms,
-    ) -> Result<Lines> {
-        let Self {
-            word,
-            extension,
-            dir,
-        } = self;
-
-        let word = Word::new(word)?;
-
-        let lang = match get_language_by_ext(&extension) {
-            Ok(lang) => lang,
-            Err(_) => {
-                // Search the occurrences if no language detected.
-                let occurrences = find_occurrence_matches_by_ext(&word, &extension, &dir).await?;
-                let (lines, indices): (Vec<String>, Vec<Vec<usize>>) = occurrences
-                    .into_par_iter()
-                    .filter_map(|line| {
-                        exact_or_inverse_terms.check_jump_line(line.build_jump_line("refs", &word))
-                    })
-                    .unzip();
-                return Ok(Lines::new(lines, indices));
-            }
-        };
-
-        let comments = get_comments_by_ext(&extension);
-
-        // render the results in group.
-        if classify {
-            let res = definitions_and_references(lang, &word, &dir, comments).await?;
-
-            let (lines, indices): (Vec<String>, Vec<Vec<usize>>) = res
-                .into_par_iter()
-                .flat_map(|(match_kind, matches)| render(matches, &match_kind, &word))
-                .unzip();
-
-            Ok(Lines::new(lines, indices))
-        } else {
-            definitions_and_references_lines(lang, &word, &dir, comments, exact_or_inverse_terms)
-                .await
-        }
-    }
-}
 
 /// Search-based jump.
 #[derive(StructOpt, Debug, Clone)]
@@ -148,15 +35,17 @@ pub struct DumbJump {
 
 impl DumbJump {
     pub async fn run(self) -> Result<()> {
-        let lang = get_language_by_ext(&self.extension)?;
-        let comments = get_comments_by_ext(&self.extension);
+        /* FIXME
+          let lang = get_language_by_ext(&self.extension)?;
+          let comments = get_comments_by_ext(&self.extension);
 
-        // TODO: also take word as query?
-        let word = Word::new(self.word)?;
+          // TODO: also take word as query?
+          let word = Word::new(self.word)?;
 
-        definitions_and_references_lines(lang, &word, &self.cmd_dir, comments, &Default::default())
-            .await?
-            .print();
+          definitions_and_references_lines(lang, &word, &self.cmd_dir, comments, &Default::default())
+              .await?
+              .print();
+        */
 
         Ok(())
     }
@@ -165,14 +54,14 @@ impl DumbJump {
         &self,
         classify: bool,
         exact_or_inverse_terms: &ExactOrInverseTerms,
-    ) -> Result<Lines> {
+    ) -> Result<UsagesInfo> {
         let searcher = RegexSearcher {
             word: self.word.to_string(),
             extension: self.extension.to_string(),
             dir: self.cmd_dir.clone(),
         };
         searcher
-            .search_references_or_occurrences(classify, exact_or_inverse_terms)
+            .search_usages(classify, exact_or_inverse_terms)
             .await
     }
 }
