+++ crates/maple_cli/src/dumb_analyzer/mod.rs
@@ -1,6 +1,3 @@
 mod find_usages;
 
-pub use self::find_usages::{
-    definitions_and_references, definitions_and_references_lines, find_occurrence_matches_by_ext,
-    get_comments_by_ext, get_language_by_ext, DefinitionRules, MatchKind,
-};
+pub use self::find_usages::{RegexSearcher, UsagesInfo};
